Move surveyor into surveyors, also, uh, remove the creepy exact duplicate of Callable that was lying around
